+++ aws/static-website/modules/cloudfront/main.tf
@@ -22,7 +22,7 @@ resource "aws_cloudfront_distribution" "this" {
     origin_id                = local.s3_origin_id
   }
 
-  aliases = [var.dns_name]
+  aliases = ["${var.dns_name}.${var.domain_name}"]
 
   enabled         = true
   is_ipv6_enabled = true