+++ aws/static-website/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## [0.1.17]() (2024-12-23)
+### Features
+* Update aws_cloudfront_distribution.aliases.
+* Add variables `domain_name` to root module and sub-module `cloudfront`.
+* Add variables `enabled_read_write_policy`, `enabled_read_only_policy` to root module and to s3 module.
+
 ## [0.1.4]() (2024-12-05)
 ### Features
 * Update terraform version constraint from `~> 1.9.8` to `>= 1.9.8`